# docs/swagger.yaml
basePath: /api/v1
definitions:
  operate.ModelError:
    properties:
      error_message:
        type: string
    type: object
  request.CreateActor:
    properties:
      birthday:
        example: 12.02.2002
        format: date
        type: string
      name:
        example: Тимоти Шаламе
        type: string
      sex:
        enum:
        - male
        - female
        example: male
        type: string
    type: object
...
  request.UpdateActor:
    properties:
      birthday:
        example: 12.02.2002
        format: date
        type: string
      name:
        example: Тимоти Шаламе
        type: string
      sex:
        enum:
        - male
        - female
        example: male
        type: string
    type: object
...
  response.Actor:
    properties:
      birthday:
        example: 12.02.2002
        format: date
        type: string
      id:
        example: 5
        format: uint64
        type: integer
      name:
        example: Тимоти Шаламе
        type: string
      sex:
        enum:
        - male
        - female
        example: male
        type: string
    type: object
  response.ActorFilms:
    properties:
      data_publish:
        example: 12.02.2023
        format: date
        type: string
      description:
        example: Futuristic film
        type: string
      id:
        example: 5
        format: uint64
        type: integer
      name:
        example: Dune
        type: string
      rating:
        example: 9
        format: uint8
        type: integer
    type: object
  response.ActorWithFilms:
    properties:
      birthday:
        example: 12.02.2002
        format: date
        type: string
      films:
        items:
          $ref: '#/definitions/response.ActorFilms'
        type: array
      id:
        example: 5
        format: uint64
        type: integer
      name:
        example: Тимоти Шаламе
        type: string
      sex:
        enum:
        - male
        - female
        example: male
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: vet_v2002@mail.ru
    name: Artem Vetoshkin
  description: Rest API Для управления базо данных фильмов
  title: Фильмотека
  version: "1.0"
paths:
  /actor:
    post:
      consumes:
      - application/json
      description: Добавляет актёра включая его имя, пол и дату рождения.
      parameters:
      - description: Информация о добавляемом актёре
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateActor'
      produces:
      - application/json
      responses:
        "201":
          description: Актёр успешно добавлен в базу
          schema:
            $ref: '#/definitions/response.Actor'
        "400":
          description: В теле запроса ошибка
          schema:
            $ref: '#/definitions/operate.ModelError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "403":
          description: У пользователя нет прав на создание актёра
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Добавление актёра.
      tags:
      - actor
  /actor/{actor_id}:
    delete:
      description: Удаляет информацию об актёре по его id.
      parameters:
      - description: Уникальный идентификатор актёра
        in: path
        name: actor_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Актёр успешно удалён
        "400":
          description: В теле запросе ошибка
          schema:
            $ref: '#/definitions/operate.ModelError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "403":
          description: У пользователя нет прав на удаление актёра
          schema:
            $ref: '#/definitions/operate.ModelError'
        "404":
          description: Актёр с указанным id не найден
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Удаление актёра.
      tags:
      - actor
    put:
      consumes:
      - application/json
      description: Обновляет данные об актёре. Все переданные поля будут обновлены.
        Отсутствующие поля будут оставлены без изменений.
      parameters:
      - description: Уникальный идентификатор актёра
        in: path
        name: actor_id
        required: true
        type: integer
      - description: Информация об обновлении
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateActor'
      produces:
      - application/json
      responses:
        "200":
          description: Актёр успешно обновлен в базе
          schema:
            $ref: '#/definitions/response.ActorWithFilms'
        "400":
          description: В теле запроса ошибка
          schema:
            $ref: '#/definitions/operate.ModelError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "403":
          description: У пользователя нет прав на обновление актёра
          schema:
            $ref: '#/definitions/operate.ModelError'
        "404":
          description: Актёр с указанным id не найден
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Обновление данных об актёре.
      tags:
      - actor
  /actor/list:
    get:
      description: Формирует список всех актёров в системы.
      produces:
      - application/json
      responses:
        "200":
          description: Список актёров успешно сформирован
          schema:
            items:
              $ref: '#/definitions/response.ActorWithFilms'
            type: array
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Получение списка актёров.
      tags:
      - actor
...
schemes:
- http
securityDefinitions:
  sessionCookie:
    description: Уникальный идентификационный номер сессии
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"